[FIX] emc_taxshelter_report: 't' in mail template
This fix weird 't' appearing in the translation of emails. There was a
mix of space and tab indentation in xml files. Tabulation indentation
generate \t code in fr_BE.po.

So this commit removes the tabulation indetation for xml file and fix
the translation in i18n/fr_BE.po.