cooperator: Scaffolding for multi-company e-mail templates
In this commit, the confirmation e-mail template (chosen at random) is
configured on the res.company model instead of fetched via its XMLID. In
a future commit, all the other cooperator-related e-mail templates will
be configured in the same way, using the common scaffolding from this
commit.

The commit introduces a bug/regression, which is that the confirmation
e-mail won't actually be sent. This is a result of the new scaffolding.
Before, this happened:

1. cooperator is marked installed.
2. The mail template data file gets sourced, creating the confirmation
   template's record.
3. The subscription request demo data file gets sourced, creating a
   subscription.request record.
4. While creating the record, an e-mail using the confirmation template
   gets sent.

In the new logic, this happens:

1. cooperator is marked installed.
2. The mail template data file gets sourced, creating the confirmation
   template's record, but NOT setting it on company 1.
3. The subscription request demo data file gets sourced, creating a
   subscription.request record.
4. While creating the record, Odoo tries to send a mail of the template
   company_id.cooperator_confirmation_mail_template, but the field is
   empty.
5. Error, obviously, but we've commented out the erroneous code, so
   let's proceed.
6. As a post-init hook, a copy of the confirmation template gets set on
   all companies' cooperator_confirmation_mail_template field.

We'll need to re-jig the logic to circumvent the error. A problem for
future me. Or you, if you're reading this in the distant future. Hello.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>