[REF] cooperator: Send confirmation e-mail as a separate step from subscription request creation
This was rather bad design to start with, and caused a bug during the
installation of this module (at which stage the mail templates were not
yet correctly set on the companies).

An additional state 'confirmed' is created. The use case is to
immediately click on it after all the values are correct.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>