cooperator: A partner's user may not have an identical email address
Using the old method, it was possible (but unlikely) that a different
partner's user was affected by changes on the partner. So instead of
matching on the login address, match on the partner_id field.

As a result of this, the code had to be adjusted to handle multiple user
records.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>